Move Live OS image detection to source
Source is responsible for ImagePath so it should be responsible also
about it's detection.
