+++ pyanaconda/modules/payload/live/live_os_interface.py
@@ -41,13 +41,6 @@ class LiveOSHandlerInterface(KickstartModuleInterfaceTemplate):
         """
         return self.implementation.space_required
 
-    def DetectLiveOSImage(self) -> Str:
-        """Try to find valid live os image.
-
-        :return: path to the base image
-        """
-        return self.implementation.detect_live_os_base_image()
-
     def SetupInstallationSourceWithTask(self) -> ObjPath:
         """Setup installation source resources."""
         return TaskContainer.to_object_path(
